+++ frontend/src/components/Profile.tsx
@@ -1,0 +1,4 @@
+
+export default function Profile(){
+    return <h2>Profile</h2>
+}
